Add Oct 29 2013 version of Mark Engelberg's suggested hash modifications
Done for integers, vectors (plus sequences and queues), sets, and
maps.  Records are included automatically when APersistentMap is
updated, because emit-defrecord causes record implementations to call
APersistentMap/mapHasheq.

There is a little extra debug printing in this commit that definitely
does not belong in Clojure.  It is there to help me track down all
calls to Util.hash(), which was causing some tests with Clojure case
statements to fail before I changed the compiler to use Util.hasheq()
instead.